Convert UserForm to TS
